+++ integrations/csharp/Oicana.Test/E2eTests.cs
@@ -5,7 +5,7 @@ using Oicana.Inputs;
 using Oicana.Interop;
 using CompilationMode = Oicana.Config.CompilationMode;
 using CompilationOptions = Oicana.Config.CompilationOptions;
-using ExportOptions = Oicana.Config.ExportOptions;
+using ExportFormat = Oicana.Config.ExportFormat;
 
 namespace Oicana.Test;
 
@@ -25,7 +25,7 @@ public class E2ETests
     {
         var template = new Template(_templateFile);
 
-        var document = template.Compile(new Dictionary<string, JsonNode>(), new Dictionary<string, BlobInput>(), ExportOptions.Png(1.0f), new CompilationOptions(CompilationMode.Development));
+        var document = template.Compile(new Dictionary<string, JsonNode>(), new Dictionary<string, BlobInput>(), ExportFormat.Png(1.0f), new CompilationOptions(CompilationMode.Development));
         using var fileStream = File.Create("e2e/development.png");
         document.CopyTo(fileStream);
     }
@@ -53,7 +53,7 @@ public class E2ETests
         {
             ["development-json"] = JsonSerializer.Deserialize<JsonNode>("{ \"name\": \"Input\", \"foo\": [41, \"testing\"] }")!
         };
-        var document = template.Compile(jsonInputs, blobInputs, ExportOptions.Png(1.0f), new CompilationOptions(CompilationMode.Production));
+        var document = template.Compile(jsonInputs, blobInputs, ExportFormat.Png(1.0f), new CompilationOptions(CompilationMode.Production));
         using var fileStream = File.Create("e2e/production.png");
         document.CopyTo(fileStream);
     }
@@ -107,7 +107,7 @@ public class E2ETests
             ["both-json"] = jsonData
         };
 
-        var document = template.Compile(jsonInputs, blobInputs, ExportOptions.Png(1.0f), new CompilationOptions(CompilationMode.Production));
+        var document = template.Compile(jsonInputs, blobInputs, ExportFormat.Png(1.0f), new CompilationOptions(CompilationMode.Production));
         using var fileStream = File.Create("e2e/all-inputs.png");
         document.CopyTo(fileStream);
     }
@@ -116,7 +116,7 @@ public class E2ETests
     public void GetsReadableErrors()
     {
         var template = new Template(_templateFile);
-        Action act = () => template.Compile(new Dictionary<string, JsonNode>(), new Dictionary<string, BlobInput>(), ExportOptions.Png(1.0f), new CompilationOptions(CompilationMode.Production));
+        Action act = () => template.Compile(new Dictionary<string, JsonNode>(), new Dictionary<string, BlobInput>(), ExportFormat.Png(1.0f), new CompilationOptions(CompilationMode.Production));
 
         act.Should()
             .Throw<OicanaException>()